Add convention for name of dvSwitch
There is the feature 'add compute-vmware nodes' which provides
extending an environment after deployment by add new compute-vmware
node and assign new Cluster to it. This feature causes the trouble,
because despite we can add new Cluster on VMware tab we can't set
cluster-to-dvSwitch mapping on plugins UI settings.

Instead that we introduce new convention for names. A name of dvSwitch
that assigned to a Cluster should be the same as name of that Cluster.

Change-Id: I42468b5f09a40287403d13745e94fb317892ef0d
Closes-bug: #1564795